# [WIP] Optimize evaluation for float arguments