@eendebakpt 2022-05-31
@oscarbenjamin This PR still has a number of failing tests, some help would be appreciated. One example I picked out: 
```
from sympy import DiracDelta
r=DiracDelta(5.1)
print(r)
```
returns `DiracDelta(5.1)` (should be `0`). 

Originally the value 5.1 was handled in `eval` (https://github.com/sympy/sympy/blob/master/sympy/functions/special/delta_functions.py#L218 I think). In this PR `eval` is skipped, but there is no key for `DiracDelta` in `sympy.core.evalf`.