@eendebakpt 2022-06-04
I checked a few more failing tests. One of them is
```
from sympy import sin, cos, re, im
n = sin(1)**2 + cos(1)**2 - 1
w=n.n(2)
v=re(w)
print(v) # should be -0.e-110, but in this PR it is -1.e-110
```

The  problem is in the `re` function. The `eval` is not executed any more, hence `arg.is_extended_real` is never triggered.
```
    @classmethod
    def eval(cls, arg):
        if arg is S.NaN:
            return S.NaN
        elif arg is S.ComplexInfinity:
            return S.NaN
        elif arg.is_extended_real: # <------
            return arg
```