@eendebakpt 2022-06-04
@oscarbenjamin It seems that skipping the `eval` part, even though the arguments are of Float type (the actual check is with `_should_evalf`) is not valid. Unless you have other suggestions I will close this PR.