@oscarbenjamin 2022-06-04
Seems like a few problems that could be improved. The DiracDelta and SchurNumber cases suggest that `eval` should be tried if the numerical evaluation fails.

The problem with `re` is I guess a bug in numerical evaluation or Float handling of `_prec`.