@eendebakpt 2022-06-04
> Seems like a few problems that could be improved. The DiracDelta and SchurNumber cases suggest that `eval` should be tried if the numerical evaluation fails.
> 

How would we check for failure? Do you mean check the output of `result.evalf(prec_to_dps(pr))`, e.g.
```
...
        if evaluate and len(args)>0:
             args =[sympify(a)for a in args]
             should = [cls._should_evalf(a) for a in args]
             if not any(p <= 0 for p in should):
                 pr = max(should)
                 result = super().__new__(cls, *args, evaluate=False, **options)
                 evaluated_result = result.evalf(prec_to_dps(pr))
                 if isinstance(evaluated_result, (Float,Add)): # or an Add of two floats
                     return evaluated_result
                 else:
                     # could not evaluate, go to full evaluation
                     result = super().__new__(cls, *args, evaluate=evaluate, **options)
                     evaluated_result = result.evalf(prec_to_dps(pr))
                     return evaluated_result
....
```