@oscarbenjamin 2022-06-04
Yes, it could be something like that. Another possibility is to check if the result returned from evalf still has the same type (rather than evaluating to something else).