@eendebakpt 2022-06-05
> Seems like a few problems that could be improved. The DiracDelta and SchurNumber cases suggest that `eval` should be tried if the numerical evaluation fails.
> 
> The problem with `re` is I guess a bug in numerical evaluation or Float handling of `_prec`.

Indeed fixes the `DiracDelta` issue. But it is not very elegant as we then have to retry and do the float evaluation again.