@eendebakpt 2022-06-06
@oscarbenjamin  On the first message in this PR I added an overview of the experiments performed. My preference at this point is to continue with the third option. I created PR https://github.com/sympy/sympy/pull/23582 as an implementation. The PR still needs some improvements (documentation, tests), but I would like to have some feedback on the following items:

- The PR passes all tests (except for one unrelated test). But the behaviour of the `Function` constructor has changed, so we cannot be sure the behaviour of sympy in all cases is unchanged. Is this acceptable?
- Performance of evaluation of float arguments improves significantly for many functions (see first message of the PR). The evaluation of exact or symbolic expressions a bit slower. I tested with
```
from sympy import cos
from sympy.core.singleton import S

w=S.Pi/4
def go():
    cos.__new__.cache_clear()
    return cos(w)
%timeit go()
# master: 18 µs ± 258 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
# PR: 19.7 µs ± 490 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```
The performance is a bit lower because in this PR for `cos(pi/4)` we evaluate `should = [cls._should_evalf(a) for a in args]`, which is not required in the master branch.  The performance seems ok, since in practice `cos(pi/4)` will be cached, and in #23587, #23586 performance is improved so combined the performance is also better for exact and symbolic arguments.
In the end, this is a trade-off between the performance on symbolic expressions and float-like expressions.
     
- Should we make the new attribute `_skip_eval_float_arguments` public? (and perhaps give it a better name)