@oscarbenjamin 2022-06-06
> My preference at this point is to continue with the third option.

That's probably best in terms of backwards compatibility. Does it also improve the problem with symbolic evaluation?