@asmeurer 2022-06-06
I don't see the need for a new `skip_eval_for_float` attribute. `_should_evalf` should be sufficient.

> The performance is a bit lower because in this PR for cos(pi/4) we evaluate should = [cls._should_evalf(a) for a in args], which is not required in the master branch.

There's basically two ways that we can do this. Either we first call `eval` and only check if `_should_evalf` if it fails (what we currently do), or we call `_should_evalf` first only call `eval` if it returns False. One optimizes for the floating point case and one optimizes for the non-floating point case. In practice, I would say the most common case is not to have floating point inputs, so it may make more sense to keep it as it is. That also has the advantage of not breaking existing code that doesn't implement `_should_evalf` or that implements it in a way that doesn't match what `eval` would do. Although I expect very little user code overrides `_should_evalf` since isn't not even documented, and we don't even override it anywhere in SymPy.